Add signup with onboarding
Now a user can be onboarded in case they have not yet created a profile.
The user goes through the onboarding screen before the Sign In With Solana signup
screen. On clicking signin with Solana the app sends a request to the chosen wallet
and then performs success ro error checks before routing the user to the dashboard.