Create live updates demo
This creates a notification for live updates using the eventsource API
and progressively updates the notification new data until a long running task
is completed.

The native code handles the eventsource and syncs it to compose state flow.
The icons are ignored for now due to the eventstream terminating mid stream.